# backend/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Roster API",
    "version": "1.0.0",
    "description": "API documentation for the Roster management system"
  },
  "servers": [
    {
      "url": "https://roster1.sigvitas.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api/check-login": {
      "post": {
        "summary": "Check employee or admin login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/save-employee-details": {
      "post": {
        "summary": "Save employee details from NewUsersLoginModel to LoginModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveEmployeeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee details saved and email sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveEmployeeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/save-new-employee-details": {
      "post": {
        "summary": "Create a new employee in NewUsersLoginModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewEmployeeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee details saved and email sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewEmployeeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "user@example.com",
            "description": "User ID (email)"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "userType": {
            "type": "string",
            "enum": ["admin", "employee"],
            "example": "employee",
            "description": "User role: admin or employee"
          }
        },
        "required": ["userId", "password", "userType"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "SaveEmployeeRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": ["email"]
      },
      "SaveEmployeeResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Employee details saved and email sent successfully"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "NewEmployeeRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Jane"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "newuser@example.com"
          },
          "contact": {
            "type": "string",
            "example": "+1234567890"
          },
          "userType": {
            "type": "string",
            "enum": ["admin", "employee"],
            "example": "employee"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": ["firstName", "lastName", "email", "contact", "userType", "password"]
      },
      "NewEmployeeResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Employee details saved and email sent successfully"
          },
          "user": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "1"
              },
              "firstName": {
                "type": "string",
                "example": "Jane"
              },
              "lastName": {
                "type": "string",
                "example": "Doe"
              },
              "email": {
                "type": "string",
                "example": "newuser@example.com"
              },
              "userId": {
                "type": "string",
                "example": "newuser@example.com"
              },
              "contact": {
                "type": "string",
                "example": "+1234567890"
              },
              "userType": {
                "type": "string",
                "example": "employee"
              }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "1"
          },
          "slNo": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "organization": {
            "type": "string",
            "example": "Example Corp"
          },
          "addressLine1": {
            "type": "string",
            "example": "123 Main St"
          },
          "addressLine2": {
            "type": "string",
            "example": "Apt 4B"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "state": {
            "type": "string",
            "example": "NY"
          },
          "country": {
            "type": "string",
            "example": "USA"
          },
          "zipcode": {
            "type": "string",
            "example": "10001"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+1234567890"
          },
          "regCode": {
            "type": "string",
            "example": "REG123"
          },
          "agentAttorney": {
            "type": "string",
            "example": "Attorney"
          },
          "dateOfPatent": {
            "type": "string",
            "example": "01-01-2020"
          },
          "agentLicensed": {
            "type": "string",
            "example": "01-01-2020"
          },
          "firmOrOrganization": {
            "type": "string",
            "example": "Example Firm"
          },
          "updatedPhoneNumber": {
            "type": "string",
            "example": "+1234567890"
          },
          "emailAddress": {
            "type": "string",
            "example": "user@example.com"
          },
          "updatedOrganization": {
            "type": "string",
            "example": "Updated Corp"
          },
          "firmUrl": {
            "type": "string",
            "example": "https://example.com"
          },
          "updatedAddress": {
            "type": "string",
            "example": "456 Main St"
          },
          "updatedCity": {
            "type": "string",
            "example": "Boston"
          },
          "updatedState": {
            "type": "string",
            "example": "MA"
          },
          "updatedCountry": {
            "type": "string",
            "example": "USA"
          },
          "updatedZipcode": {
            "type": "string",
            "example": "02108"
          },
          "linkedInProfile": {
            "type": "string",
            "example": "https://linkedin.com/in/johndoe"
          },
          "notes": {
            "type": "string",
            "example": "Additional notes"
          },
          "initials": {
            "type": "string",
            "example": "JD"
          },
          "dataUpdatedAsOn": {
            "type": "string",
            "example": "01-01-2023"
          },
          "userId": {
            "type": "string",
            "example": "user@example.com"
          },
          "admin": {
            "type": "string",
            "example": "Admin Name"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Invalid input"
          }
        }
      },
      "UnauthorizedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid credentials"
          }
        }
      },
      "NotFoundResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Resource not found"
          }
        }
      },
      "ServerErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "An error occurred"
          },
          "details": {
            "type": "string",
            "example": "Error message"
          }
        }
      }
    }
  }
}